Split html
