+++ frictionless/plugins/html/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import HtmlPlugin
+from .dialect import HtmlDialect
+from .parser import HtmlParser
