+++ frictionless/plugins/html/parser.py
@@ -1,85 +1,7 @@
 import tempfile
-from ..metadata import Metadata
-from ..dialect import Dialect
-from ..plugin import Plugin
-from ..parser import Parser
-from ..system import system
-from .. import helpers
-
-
-# Plugin
-
-
-class HtmlPlugin(Plugin):
-    """Plugin for HTML
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.html import HtmlPlugin`
-
-    """
-
-    code = "html"
-    status = "experimental"
-
-    def create_dialect(self, resource, *, descriptor):
-        if resource.format == "html":
-            return HtmlDialect(descriptor)
-
-    def create_parser(self, resource):
-        if resource.format == "html":
-            return HtmlParser(resource)
-
-
-# Dialect
-
-
-class HtmlDialect(Dialect):
-    """Html dialect representation
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.html import HtmlDialect`
-
-    Parameters:
-        descriptor? (str|dict): descriptor
-        selector? (str): HTML selector
-
-    Raises:
-        FrictionlessException: raise any error that occurs during the process
-
-    """
-
-    def __init__(self, descriptor=None, *, selector=None):
-        self.setinitial("selector", selector)
-        super().__init__(descriptor)
-
-    @Metadata.property
-    def selector(self):
-        """
-        Returns:
-            str: selector
-        """
-        return self.get("selector", "table")
-
-    # Expand
-
-    def expand(self):
-        """Expand metadata"""
-        self.setdefault("selector", self.selector)
-
-    # Metadata
-
-    metadata_profile = {  # type: ignore
-        "type": "object",
-        "additionalProperties": False,
-        "properties": {
-            "selector": {"type": "string"},
-        },
-    }
-
-
-# Parser
+from ...parser import Parser
+from ...system import system
+from ... import helpers
 
 
 class HtmlParser(Parser):
